Rename commitment.rs
