+++ infrastructure/crypto/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod common;
-pub mod commitments;
+pub mod commitment;
 pub mod hex;
 pub mod keys;
 pub mod signatures;
